+++ configs/typescript.js
@@ -8,7 +8,41 @@ module.exports = {
       "@typescript-eslint",
   ],
   "rules": {
-      "@typescript-eslint/class-name-casing": "error",
+      "@typescript-eslint/naming-convention": ["error",
+        {
+          selector: "default",
+          format: ["camelCase"],
+          leadingUnderscore: "allow"
+        },
+        {
+          selector: "variable",
+          format: ["camelCase", "UPPER_CASE", "PascalCase"],
+          leadingUnderscore: "allow"
+        },
+        {
+          selector: "function",
+          format: ["camelCase"],
+          leadingUnderscore: "allow"
+        },
+        {
+          selector: ["typeLike", "typeAlias", "typeParameter"],
+          format: ["PascalCase"],
+          leadingUnderscore: "allow"
+        },
+        {
+          selector: "enumMember",
+          format: ["PascalCase", "camelCase"]
+        },
+        {
+          selector: ["typeProperty", "typeAlias", "typeParameter"],
+          format: ["PascalCase", "camelCase"],
+          leadingUnderscore: "allow"
+        },
+        {
+          selector: ["objectLiteralProperty", "objectLiteralMethod"],
+          format: ["camelCase", "UPPER_CASE", "PascalCase", "snake_case"] // snake_case is for some external APIs and libraries
+        }
+      ],
       "@typescript-eslint/member-delimiter-style": [
           "error",
           {